+++ src/device-types/HALight.cpp
@@ -1,0 +1,196 @@
+#include "HALight.h"
+#ifndef EX_ARDUINOHA_LIGHT
+
+#include "../HAMqtt.h"
+#include "../utils/HAUtils.h"
+#include "../utils/HASerializer.h"
+
+const uint8_t HALight::DefaultBrightnessScale = 255;
+
+HALight::HALight(const char* uniqueId, const Features features) :
+    HABaseDeviceType(AHATOFSTR(HAComponentLight), uniqueId),
+    _features(features),
+    _icon(nullptr),
+    _retain(false),
+    _optimistic(false),
+    _brightnessScale(DefaultBrightnessScale),
+    _currentState(false),
+    _currentBrightness(0),
+    _stateCallback(nullptr),
+    _brightnessCallback(nullptr)
+{
+
+}
+
+bool HALight::setState(const bool state, const bool force)
+{
+    if (!force && state == _currentState) {
+        return true;
+    }
+
+    if (publishState(state)) {
+        _currentState = state;
+        return true;
+    }
+
+    return false;
+}
+
+bool HALight::setBrightness(const uint8_t brightness, const bool force)
+{
+    if (!force && brightness == _currentBrightness) {
+        return true;
+    }
+
+    if (publishBrightness(brightness)) {
+        _currentBrightness = brightness;
+        return true;
+    }
+
+    return false;
+}
+
+void HALight::buildSerializer()
+{
+    if (_serializer || !uniqueId()) {
+        return;
+    }
+
+    _serializer = new HASerializer(this, 12); // 12 - max properties nb
+    _serializer->set(AHATOFSTR(HANameProperty), _name);
+    _serializer->set(AHATOFSTR(HAUniqueIdProperty), _uniqueId);
+    _serializer->set(AHATOFSTR(HAIconProperty), _icon);
+
+    if (_retain) {
+        _serializer->set(
+            AHATOFSTR(HARetainProperty),
+            &_retain,
+            HASerializer::BoolPropertyType
+        );
+    }
+
+    if (_optimistic) {
+        _serializer->set(
+            AHATOFSTR(HAOptimisticProperty),
+            &_optimistic,
+            HASerializer::BoolPropertyType
+        );
+    }
+
+    if (_features & BrightnessFeature) {
+        _serializer->topic(AHATOFSTR(HABrightnessStateTopic));
+        _serializer->topic(AHATOFSTR(HABrightnessCommandTopic));
+
+        if (_brightnessScale != DefaultBrightnessScale) {
+            _serializer->set(
+                AHATOFSTR(HABrightnessScaleProperty),
+                &_brightnessScale,
+                HASerializer::NumberP0PropertyType
+            );
+        }
+    }
+
+    _serializer->set(HASerializer::WithDevice);
+    _serializer->set(HASerializer::WithAvailability);
+    _serializer->topic(AHATOFSTR(HAStateTopic));
+    _serializer->topic(AHATOFSTR(HACommandTopic));
+}
+
+void HALight::onMqttConnected()
+{
+    if (!uniqueId()) {
+        return;
+    }
+
+    publishConfig();
+    publishAvailability();
+
+    if (!_retain) {
+        publishState(_currentState);
+        publishBrightness(_currentBrightness);
+    }
+
+    subscribeTopic(uniqueId(), AHATOFSTR(HACommandTopic));
+
+    if (_features & BrightnessFeature) {
+        subscribeTopic(uniqueId(), AHATOFSTR(HABrightnessCommandTopic));
+    }
+}
+
+void HALight::onMqttMessage(
+    const char* topic,
+    const uint8_t* payload,
+    const uint16_t length
+)
+{
+    if (HASerializer::compareDataTopics(
+        topic,
+        uniqueId(),
+        AHATOFSTR(HACommandTopic)
+    )) {
+        handleStateCommand(reinterpret_cast<const char*>(payload), length);
+    } else if (HASerializer::compareDataTopics(
+        topic,
+        uniqueId(),
+        AHATOFSTR(HABrightnessCommandTopic)
+    )) {
+        handleBrightnessCommand(reinterpret_cast<const char*>(payload), length);
+    }
+}
+
+bool HALight::publishState(const bool state)
+{
+    return publishOnDataTopic(
+        AHATOFSTR(HAStateTopic),
+        AHATOFSTR(state ? HAStateOn : HAStateOff),
+        true
+    );
+}
+
+bool HALight::publishBrightness(const uint8_t brightness)
+{
+    if (!uniqueId() || !(_features & BrightnessFeature)) {
+        return false;
+    }
+
+    uint8_t size = HAUtils::calculateNumberSize(brightness);
+    if (size == 0) {
+        return false;
+    }
+
+    char str[size + 1]; // with null terminator
+    str[size] = 0;
+    HAUtils::numberToStr(str, brightness);
+
+    return publishOnDataTopic(AHATOFSTR(HABrightnessStateTopic), str, true);
+}
+
+void HALight::handleStateCommand(const char* cmd, const uint16_t length)
+{
+    (void)cmd;
+
+    if (!_stateCallback) {
+        return;
+    }
+
+    bool state = length == strlen_P(HAStateOn);
+    _stateCallback(state, this);
+}
+
+void HALight::handleBrightnessCommand(const char* cmd, const uint16_t length)
+{
+    if (!_brightnessCallback) {
+        return;
+    }
+
+    HAUtils::Number number = HAUtils::strToNumber(cmd);
+    if (
+        number != HAUtils::NumberMax &&
+        number >= 0 &&
+        number <= _brightnessScale
+    ) {
+        _brightnessCallback(static_cast<uint8_t>(number), this);
+    }
+}
+
+#endif