+++ src/utils/HADictionary.cpp
@@ -58,6 +58,7 @@ const char HAModeProperty[] PROGMEM = {"mode"};
 const char HACommandTemplateProperty[] PROGMEM = {"cmd_tpl"};
 const char HASpeedRangeMaxProperty[] PROGMEM = {"spd_rng_max"};
 const char HASpeedRangeMinProperty[] PROGMEM = {"spd_rng_min"};
+const char HABrightnessScaleProperty[] PROGMEM = {"bri_scl"};
 
 // topics
 const char HAConfigTopic[] PROGMEM = {"config"};
@@ -68,6 +69,8 @@ const char HACommandTopic[] PROGMEM = {"cmd_t"};
 const char HAPositionTopic[] PROGMEM = {"pos_t"};
 const char HAPercentageStateTopic[] PROGMEM = {"pct_stat_t"};
 const char HAPercentageCommandTopic[] PROGMEM = {"pct_cmd_t"};
+const char HABrightnessCommandTopic[] PROGMEM = {"bri_cmd_t"};
+const char HABrightnessStateTopic[] PROGMEM = {"bri_stat_t"};
 
 // misc
 const char HAOnline[] PROGMEM = {"online"};